# seed l2 benchmark data generation and unroll ARM Neon simd