* Add RNG seed so performance variation run over run is low.
* Run bench mark over two seed, just to make sure the perf gain we see is not due to variation in the seed
* Add unroll to ARM simd, yields about 25%+ better latency
* Add parallel accumulators, yields another 70%+ speedup
* In all this PR reduced ARM simd latency by 75%

before
```
[309.61 ms 311.68 ms 314.93 ms] L2(simd)
[318.39 ms 345.18 ms 375.98 ms] L2(simd) second rng seed
```
after
```
[86.275 ms 86.387 ms 86.610 ms] L2(simd)
```

(I'm only with my M1 mac right now, will add this to the AVX code once I'm home)